+++ packages/headless/src/controllers/core/quickview/headless-core-quickview.ts
@@ -14,14 +14,14 @@ import {CoreEngine} from '../../..';
 import {HtmlApiClient} from '../../../api/search/html/html-api-client';
 import {ClientThunkExtraArguments} from '../../../app/thunk-extra-arguments';
 
-export interface QuickviewCoreProps {
+export interface QuickviewProps {
   /**
-   * The options for the `QuickviewCore` controller.
+   * The options for the `Quickview` controller.
    */
-  options: QuickviewCoreOptions;
+  options: QuickviewOptions;
 }
 
-export interface QuickviewCoreOptions {
+export interface QuickviewOptions {
   /**
    * The result to retrieve a quickview for.
    */
@@ -33,19 +33,19 @@ export interface QuickviewCoreOptions {
   maximumPreviewSize?: number;
 }
 
-export interface QuickviewCore extends Controller {
+export interface Quickview extends Controller {
   /**
    * Retrieves the preview content for the configured result.
    */
   fetchResultContent(): void;
 
   /**
-   * The state for the `QuickviewCore` controller.
+   * The state for the `Quickview` controller.
    */
-  state: QuickviewCoreState;
+  state: QuickviewState;
 }
 
-export interface QuickviewCoreState {
+export interface QuickviewState {
   /**
    * The result preview HTML content.
    *
@@ -65,18 +65,18 @@ export interface QuickviewCoreState {
 }
 
 /**
- * Creates a Quickview controller core instance.
+ * Creates a `Quickview` controller instance.
  *
  * @param engine - The headless engine.
- * @param props - The configurable `QuickviewCore` properties.
+ * @param props - The configurable `Quickview` properties.
  * @param fetchResultContentCallback - The callback to be triggered after executing fetchResultContent.
- * @returns A `QuickviewCore` controller instance.
+ * @returns A `Quickview` controller instance.
  */
-export function buildQuickviewCore(
+export function buildCoreQuickview(
   engine: CoreEngine,
-  props: QuickviewCoreProps,
+  props: QuickviewProps,
   fetchResultContentCallback?: () => void
-): QuickviewCore {
+): Quickview {
   if (!loadQuickviewReducers(engine)) {
     throw loadReducerError;
   }