+++ packages/headless/src/controllers/did-you-mean/headless-did-you-mean.ts
@@ -1,62 +1,17 @@
-import {buildController, Controller} from '../controller/headless-controller';
-import {
-  applyDidYouMeanCorrection,
-  enableDidYouMean,
-} from '../../features/did-you-mean/did-you-mean-actions';
 import {logDidYouMeanClick} from '../../features/did-you-mean/did-you-mean-analytics-actions';
 import {executeSearch} from '../../features/search/search-actions';
-import {
-  ConfigurationSection,
-  DidYouMeanSection,
-} from '../../state/state-sections';
 import {
   QueryCorrection,
   WordCorrection,
 } from '../../api/search/search/query-corrections';
-import {loadReducerError} from '../../utils/errors';
-import {configuration, didYouMean} from '../../app/reducers';
 import {SearchEngine} from '../../app/search-engine/search-engine';
+import {
+  buildCoreDidYouMean,
+  DidYouMean,
+  DidYouMeanState,
+} from '../core/did-you-mean/headless-core-did-you-mean';
 
-export type {QueryCorrection, WordCorrection};
-
-export interface DidYouMean extends Controller {
-  /**
-   * Apply query correction using the query correction, if any, currently present in the state.
-   */
-  applyCorrection(): void;
-
-  /**
-   * The state of the `DidYouMean` controller.
-   */
-  state: DidYouMeanState;
-}
-
-export interface DidYouMeanState {
-  /**
-   * The correction that was applied to the query. If no correction was applied, will default to an empty string.
-   */
-  wasCorrectedTo: string;
-  /**
-   * The original query that was performed, without any automatic correction applied.
-   */
-  originalQuery: string;
-  /**
-   * Specifies if the query was automatically corrected by Headless.
-   *
-   * This happens when there is no result returned by the API for a particular mispelling.
-   */
-  wasAutomaticallyCorrected: boolean;
-
-  /**
-   * The query correction that is currently applied by the "did you mean" module.
-   */
-  queryCorrection: QueryCorrection;
-
-  /**
-   * Specifies if there is a query correction to apply.
-   */
-  hasQueryCorrection: boolean;
-}
+export type {QueryCorrection, WordCorrection, DidYouMean, DidYouMeanState};
 
 /**
  * The DidYouMean controller is responsible for handling query corrections.
@@ -66,46 +21,19 @@ export interface DidYouMeanState {
  * @param engine - The headless engine.
  */
 export function buildDidYouMean(engine: SearchEngine): DidYouMean {
-  if (!loadDidYouMeanReducers(engine)) {
-    throw loadReducerError;
-  }
-
-  const controller = buildController(engine);
+  const controller = buildCoreDidYouMean(engine);
   const {dispatch} = engine;
-
-  dispatch(enableDidYouMean());
-
-  const getState = () => engine.state;
 
   return {
     ...controller,
 
     get state() {
-      const state = getState();
-
-      return {
-        originalQuery: state.didYouMean.originalQuery,
-        wasCorrectedTo: state.didYouMean.wasCorrectedTo,
-        wasAutomaticallyCorrected: state.didYouMean.wasAutomaticallyCorrected,
-        queryCorrection: state.didYouMean.queryCorrection,
-        hasQueryCorrection:
-          state.didYouMean.queryCorrection.correctedQuery !== '' ||
-          state.didYouMean.wasCorrectedTo !== '',
-      };
+      return controller.state;
     },
 
     applyCorrection() {
-      dispatch(
-        applyDidYouMeanCorrection(this.state.queryCorrection.correctedQuery)
-      );
+      controller.applyCorrection();
       dispatch(executeSearch(logDidYouMeanClick()));
     },
   };
-}
-
-function loadDidYouMeanReducers(
-  engine: SearchEngine<object>
-): engine is SearchEngine<ConfigurationSection & DidYouMeanSection> {
-  engine.addReducers({configuration, didYouMean});
-  return true;
 }