+++ packages/headless/src/controllers/facets/facet/headless-facet.ts
@@ -19,8 +19,6 @@ import {FacetValueState} from '../../../features/facets/facet-api/value';
 import {SearchEngine} from '../../../app/search-engine/search-engine';
 import {
   buildCoreFacet,
-  CoreFacet,
-  CoreFacetState,
   Facet,
   FacetProps,
   FacetSearch,
@@ -28,6 +26,8 @@ import {
   FacetState,
   FacetValue,
   SpecificFacetSearchResult,
+  CoreFacet,
+  CoreFacetState,
 } from '../../core/facets/facet/headless-core-facet';
 import {CoreEngine} from '../../..';
 import {