+++ packages/headless/src/controllers/index.ts
@@ -148,23 +148,12 @@ export type {
 export {buildQuerySummary} from './query-summary/headless-query-summary';
 
 export type {
-  CoreQuerySummaryState,
-  CoreQuerySummary,
-} from './core/query-summary/headless-core-query-summary';
-export {buildCoreQuerySummary} from './core/query-summary/headless-core-query-summary';
-
-export type {
   ResultListProps,
   ResultListOptions,
   ResultListState,
   ResultList,
 } from './result-list/headless-result-list';
 export {buildResultList} from './result-list/headless-result-list';
-
-export type {
-  CoreResultListOptions,
-  CoreResultListState,
-} from './core/result-list/headless-core-result-list';
 
 export type {
   InteractiveResultOptions,
@@ -190,14 +179,15 @@ export type {
   SuggestionHighlightingOptions,
   Delimiters,
 } from './search-box/headless-search-box';
+export {buildSearchBox} from './search-box/headless-search-box';
 
 export type {
   InstantResults,
   InstantResultsState,
+  InstantResultProps,
+  InstantResultOptions,
 } from './instant-results/instant-results';
 export {buildInstantResults} from './instant-results/instant-results';
-
-export {buildSearchBox} from './search-box/headless-search-box';
 
 export type {
   SortInitialState,
@@ -282,17 +272,8 @@ export type {
   SearchStatusState,
 } from './search-status/headless-search-status';
 export {buildSearchStatus} from './search-status/headless-search-status';
-export type {CoreStatusState} from './core/status/headless-core-status';
 
 export type {ErrorPayload} from './controller/error-payload';
-
-export type {
-  QuickviewCore,
-  QuickviewCoreOptions,
-  QuickviewCoreProps,
-  QuickviewCoreState,
-} from './core/quickview/headless-core-quickview';
-export {buildQuickviewCore} from './core/quickview/headless-core-quickview';
 
 export type {
   Quickview,
@@ -408,4 +389,72 @@ export type {
   CategoryFieldSuggestionsOptions,
   CategoryFieldSuggestionsProps,
 } from './field-suggestions/category-facet/headless-category-field-suggestions';
+
+export type {
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  QuerySummaryState as CoreQuerySummaryState,
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  QuerySummary as CoreQuerySummary,
+} from './core/query-summary/headless-core-query-summary';
+export {
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version. Please use `buildQuerySummary` instead.
+   */
+  buildCoreQuerySummary,
+} from './core/query-summary/headless-core-query-summary';
+
+export type {
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  ResultListOptions as CoreResultListOptions,
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  ResultListState as CoreResultListState,
+} from './core/result-list/headless-core-result-list';
+
+export type {
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  FacetManager as CoreFacetManager,
+} from './core/facet-manager/headless-core-facet-manager';
+
+export type {
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  SearchStatus as CoreStatusState,
+} from './core/status/headless-core-status';
+
+export type {
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  Quickview as QuickviewCore,
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  QuickviewOptions as QuickviewCoreOptions,
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  QuickviewProps as QuickviewCoreProps,
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  QuickviewState as QuickviewCoreState,
+} from './core/quickview/headless-core-quickview';
+export {
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version. Please use `buildQuickview` instead.
+   */
+  buildCoreQuickview as buildQuickviewCore,
+} from './core/quickview/headless-core-quickview';
+
 export {buildCategoryFieldSuggestions} from './field-suggestions/category-facet/headless-category-field-suggestions';