+++ packages/headless/src/controllers/product-listing/facet/headless-product-listing-facet-manager.ts
@@ -6,7 +6,7 @@ import {
   FacetManager,
   FacetManagerPayload,
   FacetManagerState,
-} from '../../facet-manager/headless-facet-manager';
+} from '../../core/facet-manager/headless-core-facet-manager';
 import {ProductListingSection} from '../../../state/state-sections';
 import {ProductListingEngine} from '../../../app/product-listing-engine/product-listing-engine';