+++ packages/headless/src/test/mock-insight-state.ts
@@ -2,7 +2,10 @@ import {getCaseContextInitialState} from '../features/case-context/case-context-
 import {getConfigurationInitialState} from '../features/configuration/configuration-state';
 import {getFacetOptionsInitialState} from '../features/facet-options/facet-options-state';
 import {getCategoryFacetSetInitialState} from '../features/facets/category-facet-set/category-facet-set-state';
+import {getFacetSearchSetInitialState} from '../features/facets/facet-search-set/specific/specific-facet-search-set-state';
 import {getFacetSetInitialState} from '../features/facets/facet-set/facet-set-state';
+import {getNumericFacetSetInitialState} from '../features/facets/range-facets/numeric-facet-set/numeric-facet-set-state';
+import {getDateFacetSetInitialState} from '../features/facets/range-facets/date-facet-set/date-facet-set-state';
 import {getInsightConfigurationInitialState} from '../features/insight-configuration/insight-configuration-state';
 import {getInsightInterfaceInitialState} from '../features/insight-interface/insight-interface-state';
 import {getPaginationInitialState} from '../features/pagination/pagination-state';
@@ -12,7 +15,11 @@ import {getQueryInitialState} from '../features/query/query-state';
 import {getResultPreviewInitialState} from '../features/result-preview/result-preview-state';
 import {getSearchHubInitialState} from '../features/search-hub/search-hub-state';
 import {getSearchInitialState} from '../features/search/search-state';
+import {getSortCriteriaInitialState} from '../features/sort-criteria/sort-criteria-state';
 import {InsightAppState} from '../state/insight-app-state';
+import {getDidYouMeanInitialState} from '../features/did-you-mean/did-you-mean-state';
+import {getStaticFilterSetInitialState} from '../features/static-filter-set/static-filter-set-state';
+import {getTabSetInitialState} from '../features/tab-set/tab-set-state';
 
 export function buildMockInsightState(
   config: Partial<InsightAppState> = {}
@@ -29,10 +36,17 @@ export function buildMockInsightState(
     facetSet: getFacetSetInitialState(),
     pagination: getPaginationInitialState(),
     resultPreview: getResultPreviewInitialState(),
+    facetSearchSet: getFacetSearchSetInitialState(),
+    facetOptions: getFacetOptionsInitialState(),
     querySet: getQuerySetInitialState(),
     querySuggest: getQuerySuggestSetInitialState(),
+    numericFacetSet: getNumericFacetSetInitialState(),
+    dateFacetSet: getDateFacetSetInitialState(),
     categoryFacetSet: getCategoryFacetSetInitialState(),
-    facetOptions: getFacetOptionsInitialState(),
+    sortCriteria: getSortCriteriaInitialState(),
+    didYouMean: getDidYouMeanInitialState(),
+    staticFilterSet: getStaticFilterSetInitialState(),
+    tabSet: getTabSetInitialState(),
     ...config,
   };
 }